refactor: split the dag-run popup rendering into logical units
The popup render.rs had grown to ~700 lines rendering three unrelated
popups. Move each popup's Widget impl next to its model (clear.rs,
mark.rs) and turn the trigger popup into a module:

- trigger/mod.rs     popup state and key handling
- trigger/params.rs  parsing the Airflow params schema into entries
- trigger/render.rs  popup layout (simple confirm + params table)
- trigger/table.rs   table cell contents and column sizing
- trigger/text.rs    pure text helpers (wrap, truncate, cursor window)

The shared Yes/No button row moves to app::model::popup next to
themed_button. No behavior changes; tests move with their subjects.
